Add retry logic
Added a method to figure out if we have a network timeout error.
By default, API operations are only done once. To activate retries, the
session.Retries can be set to a number greater than 1.
If the transport error is not detected to be a timeout error, a retry
is not attempted. This is important, as retrying anything other than a
clear network timeout error can be expensive.